Now am able to save new Categories and new Sites to the DB.

+++ Repast/RepastLib/PastModel.cs
@@ -5,15 +5,15 @@ namespace RepastLib
 {
         public class Category
         {
-                public long ID { get; set; }
+                public int ID { get; set; }
                 public string CatName { get; set; }
 
         }
         public class Site
         { 
 
-                public long CatID { get; set; }
-                public long ID { get; set; }
+                public int CatID { get; set; }
+                public int ID { get; set; }
                 public string SiteName { get; set; }
                 public string Password { get; set; }
         }
